Fix big Integer Test

+++ yggdrash-core/src/test/java/io/yggdrash/core/contract/CoinContractTest.java
@@ -12,7 +12,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.lang.reflect.Field;
-import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 
@@ -66,7 +65,7 @@ public class CoinContractTest {
     public void totalSupply() {
         BigInteger res = coinContract.totalSupply();
 
-        assertEquals(BigDecimal.valueOf(1000000000000L), res);
+        assertEquals(BigInteger.valueOf(1000000000000L), res);
     }
 
     @Test
@@ -74,7 +73,7 @@ public class CoinContractTest {
         String paramStr = "{\"address\" : \"c91e9d46dd4b7584f0b6348ee18277c10fd7cb94\"}";
 
         BigInteger res = coinContract.balanceOf(createParams(paramStr));
-        assertEquals(BigDecimal.valueOf(1000000000), res);
+        assertEquals(BigInteger.valueOf(1000000000), res);
     }
 
     @Test
@@ -84,7 +83,7 @@ public class CoinContractTest {
 
         BigInteger res = coinContract.allowance(createParams(paramStr));
 
-        assertEquals(BigDecimal.ZERO, res);
+        assertEquals(BigInteger.ZERO, res);
     }
 
     @Test
@@ -117,18 +116,18 @@ public class CoinContractTest {
         String paramStr2 = "{\"address\" : \"1a0cdead3d1d1dbeef848fef9053b4f0ae06db9e\"}";
         String paramStr3 = "{\"address\" : \"c91e9d46dd4b7584f0b6348ee18277c10fd7cb94\"}";
 
-        assertEquals(BigDecimal.valueOf(1000000010),
+        assertEquals(BigInteger.valueOf(1000000010),
                 coinContract.balanceOf(createParams(paramStr2)));
-        assertEquals(BigDecimal.valueOf(999999990),
+        assertEquals(BigInteger.valueOf(999999990),
                 coinContract.balanceOf(createParams(paramStr3)));
 
         // To many amount
-        param.addProperty("amount", BigDecimal.valueOf(1000000010));
+        param.addProperty("amount", BigInteger.valueOf(1000000010));
         result = coinContract.transfer(param);
         assertFalse(result.isSuccess());
 
         // Same amount
-        param.addProperty("amount", BigDecimal.valueOf(999999990));
+        param.addProperty("amount", BigInteger.valueOf(999999990));
         result = coinContract.transfer(param);
         assertTrue(result.isSuccess());
 
@@ -156,7 +155,7 @@ public class CoinContractTest {
         }
 
         assertTrue(result.isSuccess());
-        assertEquals(BigDecimal.valueOf(300), getAllowance(owner, spender));
+        assertEquals(BigInteger.valueOf(300), getAllowance(owner, spender));
         log.debug(to + ": " + getBalance(to).toString());
         log.debug(owner + ": " + getBalance(owner).toString());
         log.debug(spender + ": " + getBalance(spender).toString());
@@ -177,7 +176,7 @@ public class CoinContractTest {
         result2 = coinContract.transferFrom(transferFromObject);
         assertTrue(result2.isSuccess());
         // reset
-        assertEquals(BigDecimal.ZERO, getAllowance(owner, spender));
+        assertEquals(BigInteger.ZERO, getAllowance(owner, spender));
     }
 
     private void approveByOwner(String owner, String spender, String amount) {
@@ -198,34 +197,34 @@ public class CoinContractTest {
         String spenderParams = "{\"address\" : \"" + spender + "\"}";
         String senderParams = "{\"address\" : \"" + owner + "\"}";
 
-        assertEquals(BigDecimal.valueOf(1000000000),
+        assertEquals(BigInteger.valueOf(1000000000),
                 coinContract.balanceOf(createParams(spenderParams)));
-        assertEquals(BigDecimal.valueOf(1000000000),
+        assertEquals(BigInteger.valueOf(1000000000),
                 coinContract.balanceOf(createParams(senderParams)));
 
         String allowanceParams = "{\"owner\" : \"" + owner + "\","
                 + "\"spender\" : \"" + spender + "\"}";
 
-        assertEquals(new BigDecimal(amount), coinContract.allowance(createParams(allowanceParams)));
+        assertEquals(new BigInteger(amount), coinContract.allowance(createParams(allowanceParams)));
     }
 
     private void assertTransferFrom(String to, String owner, String spender) {
 
         String allowanceParams = "{\"owner\" : \"" + owner + "\","
                 + "\"spender\" : \"" + spender + "\"}";
-        assertEquals(BigDecimal.valueOf(300),
+        assertEquals(BigInteger.valueOf(300),
                 coinContract.allowance(createParams(allowanceParams)));
 
         String toParams = "{\"address\" : \"" + to + "\"}";
-        assertEquals(BigDecimal.valueOf(998000000700L),
+        assertEquals(BigInteger.valueOf(998000000700L),
                 coinContract.balanceOf(createParams(toParams)));
 
         String fromParams = "{\"address\" : \"" + owner + "\"}";
-        assertEquals(BigDecimal.valueOf(999999300),
+        assertEquals(BigInteger.valueOf(999999300),
                 coinContract.balanceOf(createParams(fromParams)));
 
         String spenderParams = "{\"address\" : \"" + spender + "\"}";
-        assertEquals(BigDecimal.valueOf(1000000000),
+        assertEquals(BigInteger.valueOf(1000000000),
                 coinContract.balanceOf(createParams(spenderParams)));
     }
